extract functions to don't call the iframe builder twice
